fixed project file structure, move all components to the corresponding class hierarchy
