+++ src/test/java/org/tacc/green1/model/components/FooterComponent.java
@@ -1,9 +1,9 @@
-package org.tacc.green1.model.base.components;
+package org.tacc.green1.model.components;
 
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
-import org.tacc.green1.model.AdvancedSearchPage;
 import org.tacc.green1.model.base.Component;
+import org.tacc.green1.model.pages.AdvancedSearchPage;
 import org.tacc.green1.util.XPath;
 
 public class FooterComponent extends Component implements XPath.Footer {
