+++ src/test/java/org/tacc/green1/model/components/header/cart/CartItem.java
@@ -1,13 +1,13 @@
-package org.tacc.green1.model.cart;
+package org.tacc.green1.model.components.header.cart;
 
 import org.openqa.selenium.SearchContext;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
-import org.tacc.green1.model.base.Modal;
+import org.tacc.green1.model.base.Component;
 import org.tacc.green1.util.XPath;
 
 
-public class CartItem extends Modal implements XPath.CartItem {
+public class CartItem extends Component implements XPath.CartItem {
     @FindBy(xpath = DELETE_BUTTON)
     private WebElement deleteButton;
 
@@ -29,11 +29,9 @@ public class CartItem extends Modal implements XPath.CartItem {
     @FindBy(xpath = PRICE_TEXT)
     private WebElement priceText;
 
-
     public CartItem(SearchContext context) {
         super(context);
     }
-
 
     public CartItem setQuantity(int quantity) {
         quantityInput.clear();
