+++ src/test/java/org/tacc/green1/model/components/header/menu/HeaderMenu.java
@@ -1,13 +1,13 @@
-package org.tacc.green1.model.menu;
+package org.tacc.green1.model.components.header.menu;
 
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
-import org.tacc.green1.model.base.Modal;
+import org.tacc.green1.model.base.Component;
 import org.tacc.green1.util.DriverManager;
 import org.tacc.green1.util.XPath;
 
 
-public class HeaderMenu extends Modal implements XPath.HeaderMenu {
+public class HeaderMenu extends Component implements XPath.Header.HeaderMenu {
     @FindBy(xpath = OPTION_WOMEN_ARROW)
     private WebElement womenCategoryArrow;
 
@@ -16,7 +16,6 @@ public class HeaderMenu extends Modal implements XPath.HeaderMenu {
 
     @FindBy(xpath = OPTION_MEN_LINK)
     private WebElement menCategoryItem;
-
 
     public HeaderMenu() {
         timeoutByVisibility(womenCategoryArrow);
