+++ src/test/java/org/tacc/green1/model/components/header/menu/MenCategoryPopup.java
@@ -1,16 +1,15 @@
-package org.tacc.green1.model.menu;
+package org.tacc.green1.model.components.header.menu;
 
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
-import org.tacc.green1.model.base.Modal;
-import org.tacc.green1.model.catalog.CatalogPage;
+import org.tacc.green1.model.base.Component;
+import org.tacc.green1.model.pages.catalog.CatalogPage;
 import org.tacc.green1.util.XPath;
 
 
-public class MenCategoryPopup extends Modal implements XPath.HeaderMenu {
+public class MenCategoryPopup extends Component implements XPath.Header.HeaderMenu {
     @FindBy(xpath = OPTION_MEN_BOTTOMS)
     private WebElement menBottomsCategoryItem;
-
 
     public CatalogPage gotoMenBottomsCatalogPage() {
         timeoutByVisibility(menBottomsCategoryItem);
