+++ src/test/java/org/tacc/green1/model/pages/account/AccountPopup.java
@@ -1,9 +1,9 @@
-package org.tacc.green1.model.account;
+package org.tacc.green1.model.pages.account;
 
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
-import org.tacc.green1.model.MainPage;
 import org.tacc.green1.model.base.Component;
+import org.tacc.green1.model.pages.MainPage;
 import org.tacc.green1.util.XPath;
